[DONE] Completed practicum lesson 25
