+++ FourteenthBlock/class-based_comp_pr/src/components/Header.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class Header extends Component {
+  render() {
+    return (
+      <header>
+        <p>Welcome, {this.props.username}!</p>
+      </header>
+    );
+  }
+}
+
+export default Header;
