+++ FourteenthBlock/class-based_comp_pr/src/index.js
@@ -11,42 +11,8 @@ import React from "react";
 import ReactDOM from "react-dom";
 
 import './style.css';
+import App from './components/App';
 
-// #1
-function App() {
-  return (
-    <div>
-      <Header username='Sergey' />
-      <Greeting />
-    </div>
-  );
-}
+const maName = 'Sergey';
 
-// #2
-function Header(props) {
-  return (
-    <header>
-      <p>Welcome, {props.username}!</p>
-    </header>
-  );
-}
-
-// #3
-function Greeting() {
-  const date = new Date();
-  const hours = date.getHours();
-  let timeOfDay;
-
-  if (hours < 12) {
-    timeOfDay = "morning";
-  } else if (hours >= 12 && hours < 17) {
-    timeOfDay = "afternoon";
-  } else {
-    timeOfDay = "night";
-  }
-  return (
-    <h1>Good {timeOfDay} to you, sir or madam!</h1>
-  );
-}
-
-ReactDOM.render(<App />, document.getElementById("root"));
+ReactDOM.render(<App username={maName} />, document.getElementById("root"));
